+++ src/Application/Importer/WeatherImporterInterface.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Application\Importer;
+
+use App\Domain\Model\WeatherReportCollection;
+
+interface WeatherImporterInterface
+{
+    public function import(): WeatherReportCollection;
+}
